Added fire, added CellData struct, updated rendering engine
Now allows for linear interpolation between two colors for cells. All available colors for each cell (256) are pregenerated on start to improve performance.

To allow for this, added a new struct called CellData that replaces the old Cells integer array. For now just holds an element type and color offset for each cell.