Updated to Godot 4.3, finally made water spread faster
We love ternary operators, don't we folks?
sign += (left ? -1 : 1); is one of my favorite lines of code I've ever written
But in short, just made water act like sand if it can, and move horizontally if it can't. The key was making water only try horizontal moves it knows it can complete, instead of having a chance to try moving in a direction it won't be allowed to move in.